languages/python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "__PROJECT_NAME__"
version = "0.1.0"
description = "__PROJECT_DESCRIPTION__"
readme = "README.md"
license = {text = "PolyForm Shield License 1.0.0"}
requires-python = ">=3.10"
authors = [
    {name = "dotbrains"},
]
keywords = ["cli", "python"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.0",
]

[project.scripts]
__PROJECT_NAME__ = "__PROJECT_NAME__.cli:main"

[project.urls]
Homepage = "https://github.com/dotbrains/__PROJECT_NAME__"
Repository = "https://github.com/dotbrains/__PROJECT_NAME__"
Issues = "https://github.com/dotbrains/__PROJECT_NAME__/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["__PROJECT_NAME__"]

[tool.coverage.report]
fail_under = 80
show_missing = true

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "RUF",   # ruff-specific
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["__PROJECT_NAME__"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
